rename folder (kebab case)
